Add PTH-660 config.
This adds support for the Intuos Pro M (2017) over USB, USB with the official drivers installed, and Bluetooth. This commit is using the CTL-4100's codepath for packet handling.

Note that https://github.com/linuxwacom/input-wacom/wiki/Device-IDs lists the PTH-660 as using a 2nd-generation Wacom USB interface, while it lists the CTL-4100 as using 3rd-generation. It's possible some incompatibility may crop up here from treating this internally as a CTL-4100. I have tested this config with a PTH-660 on all 3 included interfaces, and it seems to handle pen position and button clicks properly.